Refactor: moved Task classes to entity
