+++ src/main/java/org/opennars/entity/ProcessGoal.java
@@ -12,13 +12,14 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package org.opennars.control.concept;
+package org.opennars.entity;
 
 import java.util.ArrayList;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 import org.opennars.control.DerivationContext;
+import org.opennars.control.concept.ProcessAnticipation;
 import org.opennars.entity.BudgetValue;
 import org.opennars.entity.Concept;
 import org.opennars.entity.Sentence;
