+++ src/main/java/org/opennars/entity/ProcessQuestion.java
@@ -12,7 +12,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package org.opennars.control.concept;
+package org.opennars.entity;
 
 import java.util.List;
 import org.opennars.control.DerivationContext;
